+++ example/main.dart
@@ -42,6 +42,10 @@ void main() {
 }  
   ''');
 
+  final prices = JsonPath(r'$..price');
+
+  print('All prices in the store:');
+
   /// The following code will print:
   ///
   /// $['store']['book'][0]['price']:	8.95
@@ -49,8 +53,33 @@ void main() {
   /// $['store']['book'][2]['price']:	8.99
   /// $['store']['book'][3]['price']:	22.99
   /// $['store']['bicycle']['price']:	19.95
-  JsonPath(r'$..price')
-      .filter(json)
+  prices
+      .read(json)
+      .map((match) => '${match.path}:\t${match.value}')
+      .forEach(print);
+
+  print('\n');
+
+  final bikeColor = JsonPath(r'$.store.bicycle.color');
+
+  print('A copy of the store with repainted bike:');
+  print(bikeColor.set(json, 'blue'));
+
+  print('\n');
+
+  print('Note, that the bike in the original store is still red:');
+  bikeColor
+      .read(json)
       .map((result) => '${result.path}:\t${result.value}')
       .forEach(print);
+
+  print('\n');
+
+  print('It is also possible to modify json in place '
+      'as long as the matching value is an object or a list:');
+  final someBooks = JsonPath(r'$.store.book[::2]');
+  someBooks.read(json).forEach((match) {
+    match.value['title'] = 'Banana';
+  });
+  print(json);
 }