+++ lib/json_path.dart
@@ -2,4 +2,5 @@
 library json_path;
 
 export 'package:json_path/src/json_path.dart';
-export 'package:json_path/src/result.dart';
+export 'package:json_path/src/json_path_match.dart';
+export 'package:json_path/src/predicate.dart';