+++ lib/src/selector/filter.dart
@@ -1,20 +1,24 @@
 import 'package:json_path/json_path.dart';
+import 'package:json_path/src/json_path_match.dart';
 import 'package:json_path/src/predicate.dart';
-import 'package:json_path/src/result.dart';
 import 'package:json_path/src/selector/selector.dart';
 import 'package:json_path/src/selector/selector_mixin.dart';
 
-class Filter with SelectorMixin {
-  Filter(this.name, this.predicate);
+class Filter with SelectorMixin implements Selector {
+  Filter(this.name, this.isApplicable);
 
   final String name;
 
-  final Predicate predicate;
+  final Predicate isApplicable;
 
   @override
-  Iterable<Result> filter(Iterable<Result> results) =>
-      results.where((r) => predicate(r.value));
+  Iterable<JsonPathMatch> read(Iterable<JsonPathMatch> matches) =>
+      matches.where((r) => isApplicable(r.value));
 
   @override
-  String expression([Selector previous]) => '[?$name]';
+  String expression() => '[?$name]';
+
+  @override
+  dynamic replace(dynamic json, Replacement replacement) =>
+      isApplicable(json) ? replacement(json) : json;
 }