+++ lib/src/selector/object_wildcard.dart
@@ -1,23 +1,30 @@
-import 'package:json_path/src/quote.dart';
-import 'package:json_path/src/result.dart';
-import 'package:json_path/src/selector/recursive.dart';
+import 'package:json_path/src/json_path_match.dart';
+import 'package:json_path/src/selector/quote.dart';
 import 'package:json_path/src/selector/selector.dart';
 import 'package:json_path/src/selector/selector_mixin.dart';
 
-class ObjectWildcard with SelectorMixin {
+class ObjectWildcard with SelectorMixin implements Selector {
   @override
-  Iterable<Result> filter(Iterable<Result> results) => results.map((r) {
+  Iterable<JsonPathMatch> read(Iterable<JsonPathMatch> matches) =>
+      matches.map((r) {
         if (r.value is Map) return _allProperties(r.value, r.path);
         if (r.value is List) return _allValues(r.value, r.path);
-        return <Result>[];
+        return <JsonPathMatch>[];
       }).expand((_) => _);
 
   @override
-  String expression([Selector previous]) => previous is Recursive ? '*' : '.*';
+  String expression() => '*';
 
-  Iterable<Result> _allProperties(Map map, String path) =>
-      map.entries.map((e) => Result(e.value, path + '[${Quote(e.key)}]'));
+  Iterable<JsonPathMatch> _allProperties(Map map, String path) => map.entries
+      .map((e) => JsonPathMatch(e.value, path + '[${Quote(e.key)}]'));
 
-  Iterable<Result> _allValues(List list, String path) =>
-      list.asMap().entries.map((e) => Result(e.value, path + '[${e.key}]'));
+  Iterable<JsonPathMatch> _allValues(List list, String path) => list
+      .asMap()
+      .entries
+      .map((e) => JsonPathMatch(e.value, path + '[${e.key}]'));
+
+  @override
+  dynamic replace(dynamic json, Replacement replacement) => (json is Map)
+      ? json.map((key, value) => MapEntry(key, replacement(value)))
+      : json;
 }