+++ lib/src/selector/root.dart
@@ -1,14 +1,17 @@
-import 'package:json_path/src/result.dart';
+import 'package:json_path/src/json_path_match.dart';
 import 'package:json_path/src/selector/selector.dart';
 import 'package:json_path/src/selector/selector_mixin.dart';
 
-class RootSelector with SelectorMixin {
+class RootSelector with SelectorMixin implements Selector {
   const RootSelector();
 
   @override
-  Iterable<Result> filter(Iterable<Result> results) =>
-      results.map((m) => Result(m.value, expression()));
+  Iterable<JsonPathMatch> read(Iterable<JsonPathMatch> matches) =>
+      matches.map((m) => JsonPathMatch(m.value, expression()));
 
   @override
-  String expression([Selector previous]) => r'$';
+  String expression() => r'$';
+
+  @override
+  dynamic replace(dynamic json, Replacement replacement) => replacement(json);
 }