+++ lib/src/selector/selector.dart
@@ -1,14 +1,18 @@
-import 'package:json_path/src/result.dart';
+import 'package:json_path/src/json_path_match.dart';
 
-/// Converts a set of results into a set of results
+/// Converts a set of matches into a set of matches
 abstract class Selector {
-  /// Applies this filter to the [results]
-  Iterable<Result> filter(Iterable<Result> results);
+  /// Applies this filter to the [matches]
+  Iterable<JsonPathMatch> read(Iterable<JsonPathMatch> matches);
 
   /// The filter expression as string.
-  /// The [previous] selector must be provided if being followed by this
-  String expression([Selector previous]);
+  String expression();
 
   /// Combines this expression with the [other]
   Selector then(Selector other);
+
+  /// Returns a copy of [json] with all selected values modified using [replacement] function.
+  dynamic replace(dynamic json, Replacement replacement);
 }
+
+typedef Replacement = dynamic Function(dynamic value);